% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_border}
\alias{add_border}
\title{Add border helper}
\usage{
add_border(x, geom, border_size, border_colour, bg_fill, orig)
}
\arguments{
\item{x}{magick image}

\item{geom}{Geometric shape e.g. circle, hex, heart.}

\item{border_size}{Border size in pixels.}

\item{border_colour}{Border colour}

\item{bg_fill}{Background fill}

\item{orig}{List of original dimensions e.g. `list(wd = 100, ht = 200)`}
}
\value{
Magick image
}
\description{
Add border helper
}
