% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cut_circle}
\alias{cut_circle}
\title{Circle crop helper}
\usage{
cut_circle(x, just = "center")
}
\arguments{
\item{x}{Magick images}

\item{just}{Where to justify the image prior to cropping}
}
\value{
Magick image
}
\description{
Circle crop helper
}
