% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cut_heart}
\alias{cut_heart}
\title{heart crop helper}
\usage{
cut_heart(x, just = "center")
}
\arguments{
\item{x}{Magick image}

\item{just}{Where to justify the image prior to cropping}
}
\value{
Magick images
}
\description{
heart crop helper
}
