% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cut_hex}
\alias{cut_hex}
\title{Hex crop helper}
\usage{
cut_hex(x, just = "center")
}
\arguments{
\item{x}{Magick image}

\item{just}{Where to justify the image prior to cropping}
}
\value{
Magick image
}
\description{
Hex crop helper
}
