% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cut_parallelogram}
\alias{cut_parallelogram}
\title{Parallelogram crop helper}
\usage{
cut_parallelogram(x, just = "center")
}
\arguments{
\item{x}{Magick image}

\item{just}{Where to justify the image prior to cropping}
}
\value{
Magick image
}
\description{
Parallelogram crop helper
}
