% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cut_square}
\alias{cut_square}
\title{Square crop helper}
\usage{
cut_square(x, just = "center")
}
\arguments{
\item{x}{Magick images}

\item{just}{Where to justify the image prior to}
}
\value{
Magick image
}
\description{
Square crop helper
}
