% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle-crop.R
\name{download_images}
\alias{download_images}
\title{Download images}
\usage{
download_images(images)
}
\arguments{
\item{images}{A vector of URLs to image location}
}
\value{
The paths to the downloaded images
}
\description{
If the images are a URL they will be downloaded and saved in a temporary location.
The images are cleared when the session ends
}
\examples{
\dontrun{
images <- c(
  'https://openpsychometrics.org/tests/characters/test-resources/pics/BB/1.jpg',
  'https://openpsychometrics.org/tests/characters/test-resources/pics/BB/3.jpg',
  'https://openpsychometrics.org/tests/characters/test-resources/pics/BB/9.jpg'
  )

download_images(images)
}
}
