% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{f_heart}
\alias{f_heart}
\title{heart cropping}
\usage{
f_heart(x, just = "center")
}
\arguments{
\item{x}{Magick image}

\item{just}{Where to justify the image prior to cropping}
}
\value{
Magick images
}
\description{
heart cropping
}
