% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{f_hex}
\alias{f_hex}
\title{Title}
\usage{
f_hex(x, just = "center")
}
\arguments{
\item{x}{Magick image}

\item{just}{Where to justify the image prior to cropping}
}
\value{
Magick image
}
\description{
Title
}
