% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle-crop.R
\name{is_url}
\alias{is_url}
\title{Is URL}
\usage{
is_url(path)
}
\arguments{
\item{path}{Path to image}
}
\value{
Logical vector. `TRUE` if path is a url. `FALSE` otherwise
}
\description{
Checks if the given address is a URL or not. Returns a logical vector.
}
\examples{
 is_url('https://openpsychometrics.org/tests/characters/test-resources/pics/BB/1.jpg')

 is_url("C:/path_to_not_url/image.png")
}
