% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_border}
\alias{add_border}
\title{Add border helper}
\usage{
add_border(x, geom, border_size, border_colour)
}
\arguments{
\item{x}{magick image}

\item{geom}{Geometric shape e.g. circle, hex, heart.}

\item{border_size}{Border size in pixels.}

\item{border_colour}{Border colour}
}
\value{
Magick image
}
\description{
Add border helper
}
